organize the filesystem
the build process now copies the resources folder into the exe folder,
and sets the working directory to the exe folder, so paths like
"resources/video_ship.obj" should resolve correctly ... hopefully.

works for the moment.

I'm especially concerned because the premake command for "set working
directory" is called "debugdir" which sounds like it will break in
release mode, but I guess I'll deal with that when it comes to it.